Add ability to activate/deactivate location. New location created by user must be activated by moderator.
Fixed issue with not being able to access inactive location/post via link by moderator

+++ src/API/Services/Warning/Application/Commands/Handlers/DeleteWarningCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using Domain.Repository;
+﻿using Application.Exception;
+using Domain.Repository;
 using MediatR;
 
 namespace Application.Commands.Handlers;
@@ -15,8 +16,14 @@ public class DeleteWarningCommandHandler : IRequestHandler<DeleteWarningCommand>
     public async Task<Unit> Handle(DeleteWarningCommand request, CancellationToken cancellationToken)
     {
         var warning = await _warningRepository.GetWarningAsync(request.Id);
-
-        await _warningRepository.DeleteWarningAsync(warning);
-        return Unit.Value;
+        if (warning is null)
+            throw new ItemNotFoundException("Location has not been found");
+        else if (warning.IsAuthor(request.UserId) || request.IsUserMod)
+        {
+            await _warningRepository.DeleteWarningAsync(warning);
+            return Unit.Value;
+        }
+        else
+            throw new NotAuthorizedForOperation("You are not authorized to delete this location");
     }
 }
